pyproject.toml
[tool.poetry]
name = "sentinelpost"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
pandas = "^2.0.3"
streamlit = "^1.24.1"
sqlalchemy = "^2.0.18"
pytorch-lightning = "^2.0.4"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
newsapi-python = "^0.2.7"
transformers = "^4.30.2"
pathlib = "^1.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
